feature(ERP-284): 입고 진행 상태 반환 및 진행 상태 조회 api 구현 (#46)

+++ src/main/java/org/fallguys/procurementservice/adapter/inbound/web/dto/PurchaseOrderPageResponse.java
@@ -9,7 +9,7 @@ public record PurchaseOrderPageResponse(
         int page,
         int size,
         long totalElements,
-        int totalPages,
+        long totalPages,
         boolean hasPrevious,
         boolean hasNext
 ) {
